Add code to draw the ball
the given x and y input should be the coordinates of the top left corner of the ball, although by design nothing should be drawn there, but this is more convenient and easy to understand. The FSM in the controller might fit better in datapath and another controller should be there to get x and y inputs, if we are testing, but for the game, x and y should be initiated in the middle of the screen.